add components to entry that abstract both relay and swimmer.
An entry refers to either a swimmer or a relay.  Both have
common fields, like name and age, which are frequently used
in entry or results listings.  The class now provides
methods  name, age, club and subject that apply to both
and should obviate the need to distinguish cases in views.
Here subject refers to the swimmer or the relay,
depend ing which is present.